deduplicated create edit discount form
