+++ src/pages/EditDiscount.tsx
@@ -1,10 +1,10 @@
+import { CreateEditDiscountForm } from "../components/Form/CreateEditDiscountForm";
 import CreateLayout from "../components/Layout/CreateLayout";
-import EditDiscountForm from "../components/Form/EditDiscountForm/EditDiscountForm";
 
 function EditDiscount() {
   return (
     <CreateLayout breadcrumbLabel="Modifica dati" title="Dati opportunità">
-      <EditDiscountForm />
+      <CreateEditDiscountForm />
     </CreateLayout>
   );
 }
